refactor: 파일 구조 정리 및 불필요한 객체를 상수로 변경 처음에는 ticket price 말고도 다른 정보가 주어질 것 같은 느낌이 들어 객체로 관리하였으나, 이제는 상수로 취급함
